src_code/requirements.txt
delta-spark==2.4.0
python-decouple
pyspark